# expression: implement vectorized evaluation for builtinSysDateWithFspSig