--- a/expression/builtin_time.go
+++ b/expression/builtin_time.go
@@ -1991,11 +1991,7 @@ func (b *builtinSysDateWithFspSig) evalTime(row chunk.Row) (d types.Time, isNull
 	}
 
 	loc := b.ctx.GetSessionVars().Location()
-	nowTs, err := getStmtTimestamp(b.ctx)
-	if err != nil {
-		return types.Time{}, true, err
-	}
-	now := nowTs.In(loc)
+	now := time.Now().In(loc)
 	result, err := convertTimeToMysqlTime(now, int8(fsp), types.ModeHalfEven)
 	if err != nil {
 		return types.Time{}, true, err