--- a/expression/builtin_time_test.go
+++ b/expression/builtin_time_test.go
@@ -1070,6 +1070,20 @@ func (s *testEvaluatorSuite) TestSysDate(c *C) {
 			c.Assert(times[i].String(), GreaterEqual, last.Format(types.TimeFormat))
 		}
 
+		baseFunc, _, input, output = genVecBuiltinFuncBenchCase(ctx, ast.Sysdate,
+			vecExprBenchCase{
+				retEvalType:   types.ETDatetime,
+				childrenTypes: []types.EvalType{types.ETInt},
+				geners:        []dataGenerator{&rangeInt64Gener{begin: 0, end: 7}},
+			})
+		resetStmtContext(s.ctx)
+		loc := ctx.GetSessionVars().Location()
+		startTm := time.Now().In(loc)
+		err = baseFunc.vecEvalTime(input, output)
+		c.Assert(err, IsNil)
+		for i := 0; i < 1024; i++ {
+			c.Assert(times[i].String(), GreaterEqual, startTm.Format(types.TimeFormat))
+		}
 	}
 
 	last := time.Now()