--- a/expression/builtin_time_vec_test.go
+++ b/expression/builtin_time_vec_test.go
@@ -211,11 +211,6 @@ var vecBuiltinTimeCases = map[string][]vecExprBenchCase{
 			geners:        []dataGenerator{&timeStrGener{}, &constStrGener{"%y-%m-%d"}},
 		},
 	},
-	ast.Sysdate: {
-		{retEvalType: types.ETDatetime, childrenTypes: []types.EvalType{types.ETInt},
-			geners: []dataGenerator{&rangeInt64Gener{begin: 0, end: 7}}},
-		{retEvalType: types.ETDatetime},
-	},
 	ast.GetFormat: {
 		{
 			retEvalType:   types.ETString,
@@ -228,6 +223,8 @@ var vecBuiltinTimeCases = map[string][]vecExprBenchCase{
 		// we cannot use the vectorized test framework to test builtinSysDateWithoutFspSig.
 		// We test the builtinSysDateWithoutFspSig in TestSysDate function.
 		// {retEvalType: types.ETDatetime},
+		// {retEvalType: types.ETDatetime, childrenTypes: []types.EvalType{types.ETInt},
+		// 	geners: []dataGenerator{&rangeInt64Gener{begin: 0, end: 7}}},
 	},
 }