@ekalinin 2019-11-11
/run-all-tests